# apis-service/docs/swagger.yaml
basePath: /
definitions:
  models.EndpointsParameterEntity:
    properties:
      endpointID:
        type: integer
      exampleValue:
        type: string
      id:
        type: integer
      key:
        type: string
      parameterType:
        description: e.g., Query, Header,
        type: string
      required:
        type: boolean
      valueType:
        description: e.g., string, int, boolean
        type: string
    type: object
...
  types.EndpointsDto:
    properties:
      apiID:
        type: integer
      description:
        type: string
      group:
        type: string
      groupID:
        type: integer
      methode:
        type: string
      url:
        type: string
    type: object
  types.EndpointsGroupDto:
    properties:
      apiID:
        type: integer
      description:
        type: string
      group:
        type: string
    type: object
  types.EndpointsParameterDto:
    properties:
      endpointID:
        type: integer
      exampleValue:
        type: string
      key:
        type: string
      parameterType:
        type: string
      required:
        type: boolean
      valueType:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: This is a sample server for a pet store.
  termsOfService: http://terms-of-service-url.com
  title: API Title
  version: "1.0"
paths:
...
  /endpoints:
    post:
      consumes:
      - application/json
      description: Creates a new API Endpoints
      parameters:
      - description: API Endpoints Data
        in: body
        name: endpoints
        required: true
        schema:
          $ref: '#/definitions/types.EndpointsDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an API Endpoints
      tags:
      - API Endpoints Operations
  /endpoints-group:
    get:
      description: Retrieves all Endpoints Groups
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get All Endpoints Groups
      tags:
      - Endpoints Group Operations
    post:
      consumes:
      - application/json
      description: Creates a new Endpoints Group
      parameters:
      - description: Endpoints Group Data
        in: body
        name: endpoints-group
        required: true
        schema:
          $ref: '#/definitions/types.EndpointsGroupDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an Endpoints Group
      tags:
      - Endpoints Group Operations
  /endpoints-group/:id:
    delete:
      description: Deletes a specific Endpoints Group
      parameters:
      - description: Endpoints Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an Endpoints Group
      tags:
      - Endpoints Group Operations
    patch:
      consumes:
      - application/json
      description: Updates an existing Endpoints Group
      parameters:
      - description: Endpoints Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Endpoints Group Data
        in: body
        name: endpoints-group
        required: true
        schema:
          $ref: '#/definitions/types.EndpointsGroupDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an Endpoints Group
      tags:
      - Endpoints Group Operations
  /endpoints-parameter:
    post:
      consumes:
      - application/json
      description: Creates a new Endpoints Parameter
      parameters:
      - description: Endpoints Parameter Data
        in: body
        name: endpoints-parameter
        required: true
        schema:
          $ref: '#/definitions/types.EndpointsParameterDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EndpointsParameterEntity'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an Endpoints Parameter
      tags:
      - Endpoints Parameter Operations
  /endpoints-parameter/{endpoint-id}:
    get:
      description: Retrieves all Endpoints Parameters for a specific Endpoint
      parameters:
      - description: Endpoint ID
        in: path
        name: endpoint-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EndpointsParameterEntity'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get All Endpoints Parameters
      tags:
      - Endpoints Parameter Operations
  /endpoints-parameter/{id}:
    delete:
      description: Deletes a specific Endpoints Parameter
      parameters:
      - description: Endpoints Parameter ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an Endpoints Parameter
      tags:
      - Endpoints Parameter Operations
    patch:
      consumes:
      - application/json
      description: Updates an existing Endpoints Parameter
      parameters:
      - description: Endpoints Parameter ID
        in: path
        name: id
        required: true
        type: integer
      - description: Endpoints Parameter Data
        in: body
        name: endpoints-parameter
        required: true
        schema:
          $ref: '#/definitions/types.EndpointsParameterDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EndpointsParameterEntity'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an Endpoints Parameter
      tags:
      - Endpoints Parameter Operations
  /endpoints/:api-id:
    get:
      description: Retrieves endpointss for a specific API
      parameters:
      - description: API ID
        in: path
        name: api-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get API Endpointss by API ID
      tags:
      - API Endpoints Operations
  /endpoints/:id:
    delete:
      description: Deletes a specific API Endpoints
      parameters:
      - description: Endpoints ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an API Endpoints
      tags:
      - API Endpoints Operations
    patch:
      consumes:
      - application/json
      description: Updates an existing API Endpoints
      parameters:
      - description: Endpoints ID
        in: path
        name: id
        required: true
        type: integer
      - description: API Endpoints Data
        in: body
        name: endpoints
        required: true
        schema:
          $ref: '#/definitions/types.EndpointsDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an API Endpoints
      tags:
      - API Endpoints Operations
...
schemes:
- http
- https
swagger: "2.0"